backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asteval==1.0.8
astunparse==1.6.3
backports.zstd==1.3.0
bcrypt==4.1.3
black==25.12.0
blinker==1.9.0
boto3==1.42.16
botocore==1.42.16
brightway2==2.4.7
brotli==1.2.0
bw-migrations==0.2
bw2analyzer==0.10
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.12
bw2parameters==1.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.1
ecoinvent_interface==3.1
eight==1.0.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.110.1
fasteners==0.20
flake8==7.3.0
Flask==3.1.2
flexcache==0.3
flexparser==0.4
fonttools==4.61.1
future==0.18.3
h11==0.16.0
idna==3.11
inflate64==1.0.4
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jq==1.10.0
kiwisolver==1.4.9
librt==0.7.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mrio-common-metadata==0.2.1
multivolumefile==0.2.3
mypy==1.19.1
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
peewee==3.18.3
pillow==12.1.0
Pint==0.25.2
platformdirs==4.5.1
pluggy==1.6.0
psutil==7.2.1
py7zr==1.1.0
pyasn1==0.6.1
pybcj==1.0.7
pycodestyle==2.14.0
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyecospold==4.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.3.1
pyppmd==1.3.1
PyPrind==2.11.3
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
pyxlsb==1.0.10
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
starlette==0.37.2
stats_arrays==1.0
tabulate==0.9.0
texttable==1.7.0
tqdm==4.67.1
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unicodecsv==0.14.1
Unidecode==1.4.0
urllib3==2.6.2
uvicorn==0.25.0
voluptuous==0.16.0
watchfiles==1.1.1
Werkzeug==3.1.4
Whoosh==2.7.4
wrapt==2.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
